requirements.txt
python-telegram-bot==20.8
SQLAlchemy==2.0.41
APScheduler==3.10.4
python-dotenv==1.0.1
pytest==8.2.2
